OWLookAlike: Show 100% match nicely
When the reference image comes form the data set and we do not exclude it in the Neighbours widget we show `I am 1e+2% ...` and also the number is truncated in the left TableView. This commit changes the format to always show the full number with one decimal place and also widens TableView so that 100% matches are not truncated.